requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.1
Pillow==10.1.0
matplotlib==3.8.1
scikit-image==0.22.0
python-dotenv==1.0.0
reportlab==4.0.7
werkzeug==3.0.1
